+++ solana-programs/programs/drip-v2/src/state/drip_position.rs
@@ -24,7 +24,7 @@ pub struct DripPosition {
     // This value has a default value of PairConfig.default_pair_drip_fee_bps.
     // An admin with sufficient permissions can change the position fees via the (TODO) instruction.
     // size: 8
-    pub drip_fee_bps: u64,
+    pub drip_fee_bps: u16,
     // For positions that are tokenized, this mint represents a proxy to the owner.
     // Users who have a balance for this mint and have a tokenized position is a valid "owner" for this position.
     // We store this separately and not inside the owner enum
